fixed the (values,weight) constructor
git-svn-id: https://svn.code.sf.net/p/java-ml/code/trunk@766 3162d2cb-3e1f-0410-96b0-914c58786302

+++ net/sf/javaml/core/SimpleInstance.java
@@ -89,7 +89,7 @@ public class SimpleInstance implements Instance {
         this(values, 1.0f);
     }
 
-    public SimpleInstance(double[] values, float weight) {
+    public SimpleInstance(double[] values, double weight) {
         this(values, weight, false, 0);
     }
